Factor out Util fucntions
